Add netwoorkpool testing
