+++ cmd/cl004/ac014/execute/zz_generated.flag.go
@@ -1,0 +1,13 @@
+package execute
+
+import "github.com/spf13/cobra"
+
+type flag struct {
+}
+
+func (f *flag) Init(cmd *cobra.Command) {
+}
+
+func (f *flag) Validate() error {
+	return nil
+}
