+++ cmd/cl004/ac014/explain/zz_generated.runner.go
@@ -1,0 +1,67 @@
+package explain
+
+import (
+	"context"
+	"fmt"
+	"io"
+	"strings"
+
+	"github.com/giantswarm/microerror"
+	"github.com/giantswarm/micrologger"
+	"github.com/spf13/cobra"
+
+	"github.com/giantswarm/awscnfm/v12/pkg/action"
+	"github.com/giantswarm/awscnfm/v12/pkg/action/cl004/ac014"
+	"github.com/giantswarm/awscnfm/v12/pkg/config"
+	"github.com/giantswarm/awscnfm/v12/pkg/env"
+)
+
+type runner struct {
+	flag   *flag
+	logger micrologger.Logger
+	stdout io.Writer
+	stderr io.Writer
+}
+
+func (r *runner) Run(cmd *cobra.Command, args []string) error {
+	ctx := context.Background()
+
+	err := r.flag.Validate()
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = r.run(ctx, cmd, args)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) error {
+	var err error
+
+	var e action.Explainer
+	{
+		c := ac014.ExplainerConfig{
+			Scope:         "cl004",
+			TenantCluster: config.Cluster("cl004", env.TenantCluster()),
+		}
+
+		e, err = ac014.NewExplainer(c)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	s, err := e.Explain(ctx)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	fmt.Println(strings.TrimSpace(s))
+	fmt.Println()
+
+	return nil
+}
