+++ pkg/action/cl004/ac000/plan.go
@@ -52,30 +52,35 @@ var Plan = []plan.Step{
 	{
 		Action:  "ac008",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
-		Comment: "check if kiam's pods and secrets are present",
+		Comment: "-----",
 	},
 	{
 		Action:  "ac009",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
-		Comment: "create job to test kiam aws api call",
+		Comment: "check if kiam's pods and secrets are present",
 	},
 	{
 		Action:  "ac010",
+		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
+		Comment: "create job to test kiam aws api call",
+	},
+	{
+		Action:  "ac011",
 		Backoff: plan.NewBackoff(60*time.Second, 2*time.Second),
 		Comment: "check job to test kiam aws api call",
 	},
 	{
-		Action:  "ac011",
+		Action:  "ac012",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
 		Comment: "cleanup kiam test resources",
 	},
 	{
-		Action:  "ac012",
+		Action:  "ac013",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
 		Comment: "delete cluster CRs",
 	},
 	{
-		Action:  "ac013",
+		Action:  "ac014",
 		Backoff: plan.NewBackoff(90*time.Minute, 9*time.Minute),
 		Comment: "check CRs deleted",
 	},
