+++ pkg/action/cl004/ac009/error.go
@@ -14,6 +14,16 @@ var executionFailedError = &microerror.Error{
 	Kind: "executionFailedError",
 }
 
+// jobNotCompleted is an error indicating that the job in kiam test is not jet finished
+var jobNotCompleted = &microerror.Error{
+	Kind: "jobNotCompleted",
+}
+
+// IsJobNotCompleted asserts jobNotCompleted error.
+func IsJobNotCompleted(err error) bool {
+	return microerror.Cause(err) == jobNotCompleted
+}
+
 var notFoundError = &microerror.Error{
 	Kind: "notFoundError",
 }
