+++ pkg/action/cl004/ac009/explainer.go
@@ -6,9 +6,12 @@ import (
 
 func (e *Explainer) explain(ctx context.Context) (string, error) {
 	s := `
-Test AWS API calls are working as expected. This part will create a pod (via job) 
-and try assume role "$cluster_id-Route53-Manager" and call "aws route53 list-domains" 
-to ensure kiam works as expected.
+Check if Kiam's pods and secrets are present on the tenant cluster. This test should do following:
+
+    * Ensure tls certs for Kiam are created.
+    * Ensure kiam-server and kiam-agent pods are running without errors on the tenant cluster.
+
 `
+
 	return s, nil
 }
