+++ pkg/action/cl004/ac012/executor.go
@@ -2,49 +2,8 @@ package ac012
 
 import (
 	"context"
-
-	"github.com/giantswarm/k8sclient/v4/pkg/k8sclient"
-	"github.com/giantswarm/microerror"
-	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	apiv1alpha2 "sigs.k8s.io/cluster-api/api/v1alpha2"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	pkgclient "github.com/giantswarm/awscnfm/v12/pkg/client"
-	"github.com/giantswarm/awscnfm/v12/pkg/env"
-	"github.com/giantswarm/awscnfm/v12/pkg/label"
 )
 
 func (e *Executor) execute(ctx context.Context) error {
-	var err error
-
-	var cpClients k8sclient.Interface
-	{
-		c := pkgclient.ControlPlaneConfig{
-			Logger: e.logger,
-
-			KubeConfig: env.ControlPlaneKubeConfig(),
-		}
-
-		cpClients, err = pkgclient.NewControlPlane(c)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-
-	{
-		err := cpClients.CtrlClient().DeleteAllOf(
-			ctx,
-			&apiv1alpha2.Cluster{},
-			client.InNamespace(metav1.NamespaceDefault),
-			client.MatchingLabels{label.Cluster: e.tenantCluster},
-		)
-		if errors.IsNotFound(err) {
-			// fall through
-		} else if err != nil {
-			return microerror.Mask(err)
-		}
-	}
-
 	return nil
 }
