+++ pkg/action/cl004/ac013/executor.go
@@ -3,9 +3,10 @@ package ac013
 import (
 	"context"
 
-	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/v2/pkg/apis/infrastructure/v1alpha2"
 	"github.com/giantswarm/k8sclient/v4/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	apiv1alpha2 "sigs.k8s.io/cluster-api/api/v1alpha2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -32,99 +33,16 @@ func (e *Executor) execute(ctx context.Context) error {
 	}
 
 	{
-		var list apiv1alpha2.ClusterList
-		err := cpClients.CtrlClient().List(
+		err := cpClients.CtrlClient().DeleteAllOf(
 			ctx,
-			&list,
+			&apiv1alpha2.Cluster{},
+			client.InNamespace(metav1.NamespaceDefault),
 			client.MatchingLabels{label.Cluster: e.tenantCluster},
 		)
-		if err != nil {
+		if errors.IsNotFound(err) {
+			// fall through
+		} else if err != nil {
 			return microerror.Mask(err)
-		}
-
-		if len(list.Items) != 0 {
-			return microerror.Maskf(customResourceCleanupError, "Cluster")
-		}
-	}
-
-	{
-		var list infrastructurev1alpha2.AWSClusterList
-		err := cpClients.CtrlClient().List(
-			ctx,
-			&list,
-			client.MatchingLabels{label.Cluster: e.tenantCluster},
-		)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		if len(list.Items) != 0 {
-			return microerror.Maskf(customResourceCleanupError, "AWSCluster")
-		}
-	}
-
-	{
-		var list infrastructurev1alpha2.G8sControlPlaneList
-		err := cpClients.CtrlClient().List(
-			ctx,
-			&list,
-			client.MatchingLabels{label.Cluster: e.tenantCluster},
-		)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		if len(list.Items) != 0 {
-			return microerror.Maskf(customResourceCleanupError, "G8sControlPlane")
-		}
-	}
-
-	{
-		var list infrastructurev1alpha2.AWSControlPlaneList
-		err := cpClients.CtrlClient().List(
-			ctx,
-			&list,
-			client.MatchingLabels{label.Cluster: e.tenantCluster},
-		)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		if len(list.Items) != 0 {
-			return microerror.Maskf(customResourceCleanupError, "AWSControlPlane")
-		}
-	}
-
-	{
-		var list apiv1alpha2.MachineDeploymentList
-		err := cpClients.CtrlClient().List(
-			ctx,
-			&list,
-			client.MatchingLabels{label.Cluster: e.tenantCluster},
-		)
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		if len(list.Items) != 0 {
-			return microerror.Maskf(customResourceCleanupError, "MachineDeployment")
-		}
-	}
-
-	{
-		var list infrastructurev1alpha2.AWSMachineDeploymentList
-		err := cpClients.CtrlClient().List(
-			ctx,
-			&list,
-			client.MatchingLabels{label.Cluster: e.tenantCluster},
-		)
-
-		if err != nil {
-			return microerror.Mask(err)
-		}
-
-		if len(list.Items) != 0 {
-			return microerror.Maskf(customResourceCleanupError, "AWSMachineDeployment")
 		}
 	}
